+++ src/twg2/parser/condition/ParserFactory.java
@@ -6,7 +6,7 @@ package twg2.parser.condition;
  */
 public interface ParserFactory<T extends ParserCondition> {
 
-	/**
+	/** Whether this parser factory creates compound parser conditions.
 	 * @return true if the element parsed by this condition can contain sub-elements, i.e. if other elements can be parsed from within the source of this element,
 	 * false if this condition cannot contain sub-elements
 	 */
@@ -17,5 +17,17 @@ public interface ParserFactory<T extends ParserCondition> {
 	 * @return a new {@link ParserCondition} that match against this {@code ParserFactory}
 	 */
 	public T createParser();
+
+
+	/** Return an existing {@link ParserCondition} for reuse if possible.
+	 * Implementations are not required to do anything if they so choose.<br>
+	 * Factories which wish to implement parser reuse or pooling can provide an implementation.
+	 * So it is recommended that callers always call this method to benefit from any
+	 * performance optimizations by the implementation.
+	 * @param parser the parser condition to return
+	 */
+	public default void returnParser(T parser) {
+		// no-op
+	}
 
 }