+++ src/twg2/parser/condition/text/CharParser.java
@@ -4,7 +4,7 @@ import twg2.parser.condition.ParserCondition;
 import twg2.parser.textFragment.TextFragmentRef;
 import twg2.parser.textParser.TextParser;
 
-/** A char parser, used to determine whether a series of characters match a requirement.<br>
+/** A char parser is used to determine whether a series of characters match a requirement.<br>
  * A {@code CharParser} is stateful and must keep track of previous characters passed to {@link #acceptNext(char, TextParser)}
  * and return false once the set of characters forms an invalid sequence.
  * @author TeamworkGuy2
@@ -31,6 +31,10 @@ public interface CharParser extends ParserCondition {
 	public CharParser copy();
 
 
+	/** Recycle this char parser. Resetting it back to its default state. 
+	 * Default interface method throws {@link UnsupportedOperationException}
+	 * @return This char parser recycled back to its default state
+	 */
 	@Override
 	public default CharParser recycle() {
 		throw new UnsupportedOperationException("CharParser recycling not supported");
@@ -39,34 +43,21 @@ public interface CharParser extends ParserCondition {
 
 	@Override
 	public default CharParser copyOrReuse() {
-		CharParser filter = null;
 		if(this.canRecycle()) {
-			filter = this.recycle();
+			return this.recycle();
 		}
 		else {
-			filter = this.copy();
+			return this.copy();
 		}
-		return filter;
-	}
-
-
-	/**
-	 * @see #readConditional(TextParser, StringBuilder)
-	 */
-	public default boolean readConditional(TextParser buf) {
-		return readConditional(buf, null);
 	}
 
 
 	/** Read a {@link TextParser} until it {@link #isComplete()} or until {@link #acceptNext(char, TextParser)} returns false.
-	 * The {@code dst} buffer is filled with the accepted input if a complete token is read. The input {@code buf} and {@code dst}
-	 * buffer are reset to their initial position and length if a complete token is not found.
+	 * The text parser {@code buf} is reset to its initial position if a complete token is not found.
 	 * @param buf the {@link TextParser} to read input from
-	 * @param dst the destination buffer to store matched input in
 	 * @return true if a complete token is read, false if not
 	 */
-	public default boolean readConditional(TextParser buf, StringBuilder dst) {
-		int off = dst != null ? dst.length() : 0;
+	public default boolean readConditional(TextParser buf) {
 		int count = 0;
 		while(!this.isComplete()) {
 			if(!buf.hasNext()) {
@@ -76,14 +67,35 @@ public interface CharParser extends ParserCondition {
 			char ch = buf.nextChar();
 			if(!this.acceptNext(ch, buf)) {
 				buf.unread(count);
-				if(dst != null) {
-					dst.setLength(off);
-				}
 				return false;
 			}
-			if(dst != null) {
-				dst.append(ch);
+		}
+		return true;
+	}
+
+
+	/** Read a {@link TextParser} until it {@link #isComplete()} or until {@link #acceptNext(char, TextParser)} returns false.
+	 * The {@code dst} buffer is filled with the accepted input if a complete token is read. The text parser {@code buf} and {@code dst}
+	 * buffer are reset to their initial position and length if a complete token is not found.
+	 * @param buf the {@link TextParser} to read input from
+	 * @param dst the destination buffer to store matched input in
+	 * @return true if a complete token is read, false if not
+	 */
+	public default boolean readConditional(TextParser buf, StringBuilder dst) {
+		int off = dst.length();
+		int count = 0;
+		while(!this.isComplete()) {
+			if(!buf.hasNext()) {
+				return false;
 			}
+			count++;
+			char ch = buf.nextChar();
+			if(!this.acceptNext(ch, buf)) {
+				buf.unread(count);
+				dst.setLength(off);
+				return false;
+			}
+			dst.append(ch);
 		}
 		return true;
 	}